Move semantic search to typescript
