+++ package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@remix-run/cloudflare-workers": "^1.13.0",
     "@remix-run/react": "^1.13.0",
+    "@tensorflow-models/universal-sentence-encoder": "^1.3.3",
+    "@tensorflow/tfjs": "^4.7.0",
     "copy-to-clipboard": "^3.3.3",
     "lodash": "^4.17.21",
     "markdown-it": "^13.0.1",
